moved adoption button anf form to same file
